Extract event_queue.py (smaller).
